facility updated

+++ server/src/modules/category/category.service.ts
@@ -1,11 +1,11 @@
-import {Injectable} from '@nestjs/common';
-import {CreateCategoryInput} from './dto/create-category.input';
-import {UpdateCategoryInput} from './dto/update-category.input';
-import {NodeService} from '../node/node.service';
-import {InjectModel} from '@nestjs/mongoose';
-import {Model} from 'mongoose';
-import {Category, CategoryDocument} from './entities/category.entity';
-import fs from "fs";
+import { Injectable } from '@nestjs/common';
+import { CreateCategoryInput } from './dto/create-category.input';
+import { UpdateCategoryInput } from './dto/update-category.input';
+import { NodeService } from '../node/node.service';
+import { InjectModel } from '@nestjs/mongoose';
+import { Model } from 'mongoose';
+import { Category, CategoryDocument } from './entities/category.entity';
+import fs from 'fs';
 
 @Injectable()
 export class CategoryService {
@@ -16,7 +16,8 @@ export class CategoryService {
 
   async create(createCategoryInput: CreateCategoryInput, createdBy: string) {
     createCategoryInput._id = await this.nodeService.create(
-      createCategoryInput.parent, createdBy
+      createCategoryInput.parent,
+      createdBy,
     );
     return new this.categoryModel(createCategoryInput).save();
   }
@@ -30,16 +31,12 @@ export class CategoryService {
   }
 
   update(id: string, updateCategoryInput: UpdateCategoryInput) {
-    return this.categoryModel.findByIdAndUpdate(
-      id,
-      updateCategoryInput as any
-    );
+    return this.categoryModel.findByIdAndUpdate(id, updateCategoryInput as any);
   }
 
   async remove(id: string) {
-    const path = `./uploads/${id}`
-    if (fs.existsSync(path))
-      fs.rmSync(path)
+    const path = `./uploads/${id}`;
+    if (fs.existsSync(path)) fs.rmSync(path);
     await this.nodeService.remove(id);
     return this.categoryModel.findByIdAndRemove(id);
   }
